+++ src/plugins/exec/sim-host-handlers.js
@@ -36,6 +36,7 @@ module.exports = {
                     document.getElementById('exec-service').textContent = service;
                     document.getElementById('exec-action').textContent = action;
 
+                    /*eslint-disable no-inner-declarations */
                     function handleSuccess() {
                         exec(success, true);
                     }
@@ -65,6 +66,7 @@ module.exports = {
                         sendExecUnhandledTelemetry(service, action, false, isSuccess, resultHasValue(result), shouldPersist);
                         func.apply(null, result ? [result] : []);
                     }
+                    /*eslint-enable no-inner-declarations */
 
                     // Do this each time to capture the values from the current closure. Also, use this approach rather than
                     // addEventListener(), as it can prove difficult to remove the event listener.