+++ src/plugins/exec/sim-host.js
@@ -1,20 +1,30 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 
+/*global CordovaLabeledValue: false, CordovaItem: false */
+// These globals are cordova-simulate custom elements.
+
 var savedSims = require('./saved-sims');
 var event = require('event');
+
+var emptyLabel;
+var execList;
 
 module.exports = {
     initialize: function () {
         var sims = savedSims.sims;
 
-        var execList = document.getElementById('exec-list');
+        emptyLabel = document.getElementById('empty-label');
+        execList = document.getElementById('exec-list');
         execList.addEventListener('itemremoved', function (e) {
             savedSims.removeSim(e.detail.itemIndex);
-            addEmptyItem();
+
+            if (!savedSims.sims.length) {
+                showEmptyLabel();
+            }
         });
 
         event.on('saved-sim-added', function (sim) {
-            removeEmptyItem();
+            hideEmptyLabel();
             execList.addItem(cordovaItemFromSim(sim));
         });
 
@@ -24,7 +34,7 @@ module.exports = {
             });
         } else {
             // Create a "No values saved" item
-            addEmptyItem();
+            showEmptyLabel();
         }
     }
 };
@@ -38,6 +48,7 @@ function cordovaItemFromSim(sim) {
         try {
             value = JSON.stringify(value);
         } catch (e) {
+            // ignore
         }
     }
 
@@ -47,36 +58,13 @@ function cordovaItemFromSim(sim) {
     return cordovaItem;
 }
 
-function createEmptyItem() {
-    var labeledValue = new CordovaLabeledValue();
-    labeledValue.label = 'No values saved';
-    labeledValue.value = '';
-    var cordovaItem = new CordovaItem();
-    cordovaItem.appendChild(labeledValue);
-    return cordovaItem;
+function showEmptyLabel() {
+    emptyLabel.classList.remove('cordova-hidden');
+    execList.classList.add('cordova-hidden');
 }
 
-var hasEmptyItem = false;
-function addEmptyItem() {
-    if (hasEmptyItem) {
-        return;
-    }
-
-    var execList = document.getElementById('exec-list');
-    var sims = savedSims.sims;
-    if (sims.length === 0) {
-        execList.addItem(createEmptyItem());
-        hasEmptyItem = true;
-    }
-}
-
-function removeEmptyItem() {
-    if (!hasEmptyItem) {
-        return;
-    }
-
-    var execList = document.getElementById('exec-list');
-    execList.removeItem(0);
-    hasEmptyItem = false;
+function hideEmptyLabel() {
+    emptyLabel.classList.add('cordova-hidden');
+    execList.classList.remove('cordova-hidden');
 }