+++ src/plugins/ionic-plugin-keyboard/sim-host-handlers.js
@@ -4,7 +4,7 @@ module.exports = {
     'Keyboard': {
         'init': function (successCallback) {
             // Report to the app the keyboard is hidden
-            successCallback("H");
+            successCallback('H');
         }
     }
 };