+++ src/server/socket.js
@@ -1,11 +1,21 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 
-var log = require('./utils/log'),
+/*
+ * This module implements the server part of the communication protocol as
+ * described at
+ *      https://github.com/Microsoft/cordova-simulate/wiki/The-cordova-simulate-bootstrap-protocol
+ * with a limitation that only single app-host is supported per server (and
+ * sim-host).
+ */
+
+var Q = require('q'),
+    log = require('./utils/log'),
     LiveReload = require('./live-reload/live-reload-server');
 
-var APP_HOST = 'app-host';
-var SIM_HOST = 'sim-host';
-var DEBUG_HOST = 'debug-host';
+// make variable match the literal
+var APP_HOST = 'APP_HOST',
+    SIM_HOST = 'SIM_HOST',
+    DEBUG_HOST = 'DEBUG_HOST';
 
 /**
  * @constructor
@@ -19,6 +29,14 @@ function SocketServer(simulator) {
     this._pendingEmits[SIM_HOST] = [];
     this._pendingEmits[DEBUG_HOST] = [];
 
+    /* Deferred are used to delay the actions until an event happens.
+     * Two major events are kept track of:
+     *      - When app-host connects (sends initial message)
+     *      - When sim-host notifies it's ready to serve an app-host.
+     */
+    this._whenAppHostConnected = Q.defer();
+    this._whenSimHostReady     = Q.defer();
+
     var config = this._simulator.config,
         project = this._simulator.project,
         telemetry = this._simulator.telemetry;
@@ -26,130 +44,223 @@ function SocketServer(simulator) {
     this._liveReload = new LiveReload(project, telemetry, config.forcePrepare);
 }
 
+SocketServer.prototype._resetAppHostState = function () {
+    this._whenAppHostConnected = Q.defer();
+    this._whenAppHostConnected.promise.then(this._onAppHostConnected.bind(this));
+};
+
+SocketServer.prototype._resetSimHostState = function () {
+    this._whenSimHostReady = Q.defer();
+    this._whenSimHostReady.promise.then(this._onSimHostReady.bind(this));
+};
+
+SocketServer.prototype._setupAppHostHandlers = function () {
+    log.log('Setup handlers for APP_HOST');
+
+    var config = this._simulator.config;
+
+    this._subscribeTo(APP_HOST, 'exec', function (data) {
+        this._emitTo(SIM_HOST, 'exec', data);
+    }.bind(this));
+
+    this._subscribeTo(APP_HOST, 'plugin-message', function (data) {
+        this._emitTo(SIM_HOST, 'plugin-message', data);
+    }.bind(this));
+
+    this._subscribeTo(APP_HOST, 'plugin-method', function (data, callback) {
+        this._emitTo(SIM_HOST, 'plugin-method', data, callback);
+    }.bind(this));
+
+    // Set up live reload if necessary.
+    if (config.liveReload) {
+        log.log('Starting live reload.');
+        this._liveReload.start(this._hostSockets[APP_HOST]);
+    }
+
+    // Set up telemetry if necessary.
+    if (config.telemetry) {
+        this._subscribeTo(APP_HOST, 'telemetry', function (data) {
+            this._simulator.telemetry.handleClientTelemetry(data);
+        }.bind(this));
+
+        this._emitTo(APP_HOST, 'init-telemetry');
+    }
+
+    // Set up xhr proxy
+    if (config.xhrProxy) {
+        this._emitTo(APP_HOST, 'init-xhr-proxy');
+    }
+
+    // setup touch events support
+    if (config.touchEvents) {
+        this._emitTo(APP_HOST, 'init-touch-events');
+    }
+
+    this._handlePendingEmits(APP_HOST);
+};
+
+SocketServer.prototype._handleSimHostRegistration = function () {
+    this._subscribeTo(SIM_HOST, 'ready', this._handleSimHostReady.bind(this), true);
+
+    this._emitTo(SIM_HOST, 'init');
+};
+
+SocketServer.prototype._onAppHostConnected = function () {
+    this._whenSimHostReady.promise.then(function () {
+        this._subscribeTo(APP_HOST, 'app-plugin-list', this._handleAppPluginList.bind(this), true);
+        this._emitTo(APP_HOST, 'init');
+    }.bind(this));
+};
+
+SocketServer.prototype._onSimHostReady = function () {
+    this._setupAppHostHandlers();
+};
+
+SocketServer.prototype._handleSimHostReady = function () {
+    // Resolve the deferred
+    this._whenSimHostReady.resolve();
+
+    this._setupSimHostHandlers();
+
+    this._whenAppHostConnected.promise
+        .then(this._onAppHostConnected.bind(this));
+};
+
+SocketServer.prototype._handleAppPluginList = function (data) {
+    this._whenSimHostReady.promise.then(function () {
+        this._subscribeTo(SIM_HOST, 'start', this._handleStart.bind(this), true);
+        this._emitTo(SIM_HOST, 'app-plugin-list', data);
+    }.bind(this));
+};
+
+SocketServer.prototype._handleStart = function () {
+    this._emitTo(APP_HOST, 'start');
+};
+
+SocketServer.prototype._setupSimHostHandlers = function () {
+    log.log('Setup handlers for SIM_HOST');
+
+    this._subscribeTo(SIM_HOST, 'exec-success', function (data) {
+        this._emitTo(APP_HOST, 'exec-success', data);
+    }.bind(this));
+
+    this._subscribeTo(SIM_HOST, 'exec-failure', function (data) {
+        this._emitTo(APP_HOST, 'exec-failure', data);
+    }.bind(this));
+
+    this._subscribeTo(SIM_HOST, 'plugin-message', function (data) {
+        this._emitTo(APP_HOST, 'plugin-message', data);
+    }.bind(this));
+
+    this._subscribeTo(SIM_HOST, 'plugin-method', function (data, callback) {
+        this._emitTo(APP_HOST, 'plugin-method', data, callback);
+    }.bind(this));
+
+    this._subscribeTo(SIM_HOST, 'debug-message', function (data) {
+        this._emitTo(DEBUG_HOST, data.message, data.data);
+    }.bind(this));
+
+    // Set up telemetry if necessary.
+    if (this._simulator.config.telemetry) {
+        this._subscribeTo(SIM_HOST, 'telemetry', function (data) {
+            this._simulator.telemetry.handleClientTelemetry(data);
+        }.bind(this));
+
+        this._emitTo(SIM_HOST, 'init-telemetry');
+    }
+
+    this._handlePendingEmits(SIM_HOST);
+
+};
+
 SocketServer.prototype.init = function (server) {
     var that = this,
-        telemetry = this._simulator.telemetry,
         config = this._simulator.config;
 
     this._io = require('socket.io')(server);
 
     this._io.on('connection', function (socket) {
         socket.on('register-app-host', function () {
-            log.log('App-host registered with server.');
-
-            // It only makes sense to have one app host per server. If more than one tries to connect, always take the
-            // most recent.
-            this._hostSockets[APP_HOST] = socket;
-
-            socket.on('exec', function (data) {
-                that.emitToHost(SIM_HOST, 'exec', data);
-            });
-
-            socket.on('plugin-message', function (data) {
-                that.emitToHost(SIM_HOST, 'plugin-message', data);
-            });
-
-            socket.on('plugin-method', function (data, callback) {
-                that.emitToHost(SIM_HOST, 'plugin-method', data, callback);
-            });
-
-            socket.on('telemetry', function (data) {
-                telemetry.handleClientTelemetry(data);
-            });
-
-            socket.on('debug-message', function (data) {
-                that.emitToHost(DEBUG_HOST, data.message, data.data);
-            });
-
-            // Set up live reload if necessary.
-            if (config.liveReload) {
-                log.log('Starting live reload.');
-                that._liveReload.start(socket);
+            log.log('APP_HOST connected to the server');
+            if (that._hostSockets[APP_HOST]) {
+                log.log('Overriding previously connected APP_HOST');
+                that._resetAppHostState();
             }
-
-            // Set up telemetry if necessary.
-            if (config.telemetry) {
-                socket.emit('init-telemetry');
-            }
-
-            // Set up xhr proxy
-            if (config.xhrProxy) {
-                socket.emit('init-xhr-proxy');
-            }
-
-            // setup touch events support
-            if (config.touchEvents) {
-                socket.emit('init-touch-events');
-            }
-
-            this._handlePendingEmits(APP_HOST);
-        }.bind(this));
+            that._hostSockets[APP_HOST] = socket;
+            that._whenSimHostReady.promise
+                .then(that._onSimHostReady.bind(that));
+            that._whenAppHostConnected.resolve();
+        });
 
         socket.on('register-simulation-host', function () {
-            log.log('Simulation host registered with server.');
-
-            // It only makes sense to have one simulation host per server. If more than one tries to connect, always
-            // take the most recent.
-            this._hostSockets[SIM_HOST] = socket;
-
-            socket.on('exec-success', function (data) {
-                that.emitToHost(APP_HOST, 'exec-success', data);
-            });
-            socket.on('exec-failure', function (data) {
-                that.emitToHost(APP_HOST, 'exec-failure', data);
-            });
-
-            socket.on('plugin-message', function (data) {
-                that.emitToHost(APP_HOST, 'plugin-message', data);
-            });
-
-            socket.on('plugin-method', function (data, callback) {
-                that.emitToHost(APP_HOST, 'plugin-method', data, callback);
-            });
-
-            socket.on('telemetry', function (data) {
-                telemetry.handleClientTelemetry(data);
-            });
-
-            socket.on('debug-message', function (data) {
-                that.emitToHost(DEBUG_HOST, data.message, data.data);
-            });
-
-            // Set up telemetry if necessary.
-            if (config.telemetry) {
-                socket.emit('init-telemetry');
+            log.log('SIM_HOST connected to the server');
+            if (that._hostSockets[SIM_HOST]) {
+                log.log('Overriding previously connected SIM_HOST');
+                that._resetSimHostState();
             }
-
-            this._handlePendingEmits(SIM_HOST);
-        }.bind(this));
+            that._hostSockets[SIM_HOST] = socket;
+            that._handleSimHostRegistration(socket);
+        });
 
         socket.on('register-debug-host', function (data) {
-            log.log('Debug-host registered with server.');
+            log.log('DEBUG_HOST registered with server.');
 
             // It only makes sense to have one debug host per server. If more than one tries to connect, always take
             // the most recent.
-            this._hostSockets[DEBUG_HOST] = socket;
+            that._hostSockets[DEBUG_HOST] = socket;
 
             if (data && data.handlers) {
-                socket.on('disconnect', function () {
-                    log.log('Debug-host disconnected.');
-                    config.debugHostHandlers = null;
-                });
                 config.debugHostHandlers = data.handlers;
             }
 
-            this._handlePendingEmits(DEBUG_HOST);
-        }.bind(this));
-    }.bind(this));
+            that._handlePendingEmits(DEBUG_HOST);
+        });
+
+        socket.on('disconnect', function () {
+            var type;
+            Object.keys(that._hostSockets).forEach(function (t) {
+                if (that._hostSockets[t] === socket) {
+                    type = t;
+                }
+            });
+            if (!type) {
+                log.log('Disconnect for an inactive socket');
+                return;
+            }
+            that._hostSockets[type] = undefined;
+            log.log(type + ' disconnected from the server');
+            switch (type) {
+                case APP_HOST:
+                    that._resetAppHostState();
+                    break;
+                case SIM_HOST:
+                    that._resetSimHostState();
+                    break;
+                case DEBUG_HOST:
+                    config.debugHostHandlers = null;
+                    break;
+            }
+
+        });
+
+    });
 };
 
 /**
- * @param {string} host
- * @param {string} msg
- * @param {object} data
- * @param {Function} callback
+ * @private
  */
-SocketServer.prototype.emitToHost = function (host, msg, data, callback) {
+SocketServer.prototype._handlePendingEmits = function (host) {
+    log.log('Handling pending emits for ' + host);
+    this._pendingEmits[host].forEach(function (pendingEmit) {
+        this._emitTo(host, pendingEmit.msg, pendingEmit.data, pendingEmit.callback);
+    }.bind(this));
+    this._pendingEmits[host] = [];
+};
+
+/**
+ * @private
+ */
+SocketServer.prototype._emitTo = function (host, msg, data, callback) {
     var socket = this._hostSockets[host];
     if (socket) {
         log.log('Emitting \'' + msg + '\' to ' + host);
@@ -160,14 +271,26 @@ SocketServer.prototype.emitToHost = function (host, msg, data, callback) {
     }
 };
 
-SocketServer.prototype.invalidateSimHost = function () {
-    // Simulation host is being refreshed, so we'll wait on a new connection.
-    this._hostSockets[SIM_HOST] = null;
+/**
+ * @private
+ */
+SocketServer.prototype._subscribeTo = function (host, msg, handler, once) {
+    var socket = this._hostSockets[host],
+        method = once ? 'once' : 'on';
+    if (socket) {
+        socket[method](msg, handler);
+    } else {
+        log.log('Subscribing to a disconnected ' + host + ' wanting \'' + msg + '\'');
+    }
+};
+
+SocketServer.prototype.reloadSimHost = function () {
+    this._emitTo(SIM_HOST, 'refresh');
 };
 
 SocketServer.prototype.closeConnections = function () {
     // stop watching file changes
-    if (this._simulator._config.liveReload) {
+    if (this._simulator.config.liveReload) {
         this._liveReload.stop();
     }
 
@@ -189,17 +312,6 @@ SocketServer.prototype.closeConnections = function () {
     this._pendingEmits[APP_HOST] = [];
     this._pendingEmits[SIM_HOST] = [];
     this._pendingEmits[DEBUG_HOST] = [];
-};
-
-/**
- * @param {string} host
- */
-SocketServer.prototype._handlePendingEmits = function (host) {
-    this._pendingEmits[host].forEach(function (pendingEmit) {
-        log.log('Handling pending emit \'' + pendingEmit.msg + '\' to ' + host);
-        this.emitToHost(host, pendingEmit.msg, pendingEmit.data, pendingEmit.callback);
-    }.bind(this));
-    this._pendingEmits[host] = [];
 };
 
 module.exports = SocketServer;